# openapi-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "To-Do List API",
    "version": "1.0.0",
...
    "contact": {
      "name": "API Support",
      "email": "support@todolist-api.com",
      "url": "https://github.com/your-org/todolist-api"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.todolist.com/api",
      "description": "Production server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "required": ["code", "message", "timestamp"],
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code for programmatic handling",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "The provided data is invalid"
          },
          "details": {
            "description": "Additional error details (validation errors, etc.)",
            "example": {
              "field": "name",
              "constraint": "minLength",
              "value": ""
            },
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Request timestamp when error occurred",
            "example": "2025-07-04T10:00:00.000Z"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/lists/{listId}/tasks": {
      "post": {
        "summary": "Create a new task in a list",
        "description": "Creates a new task within the specified list",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the list to add the task to"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "description": "Title of the task"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Optional description of the task"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional deadline for the task (future date)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "List not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get tasks for a specific list",
        "description": "Retrieves all tasks belonging to the specified list with optional filtering and sorting",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The list ID"
          },
          {
            "in": "query",
            "name": "includeCompleted",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Whether to include completed tasks"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": ["title", "createdAt", "updatedAt", "deadline"],
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            },
            "description": "Sort order"
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number",
                          "example": 5
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid UUID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "List not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks": {
      "get": {
        "summary": "Get all tasks with optional filters",
        "description": "Retrieves all tasks with optional filtering and sorting",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "query",
            "name": "listId",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Filter tasks by list ID"
          },
          {
            "in": "query",
            "name": "completed",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "description": "Filter tasks by completion status"
          },
          {
            "in": "query",
            "name": "includeCompleted",
            "schema": {
              "type": "boolean",
              "enum": ["title", "deadline", "createdAt", "updatedAt"],
              "default": true
            },
            "description": "Whether to include completed tasks"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": ["title", "createdAt", "updatedAt", "deadline"],
              "default": "createdAt"
            },
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "string",
              "enum": ["asc", "desc"],
              "default": "desc"
            },
            "description": "Sort order"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Number of tasks to return"
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "description": "Number of tasks to skip"
          },
          {
            "in": "query",
            "name": "deadlineFrom",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter tasks with deadline from this date"
          },
          {
            "in": "query",
            "name": "deadlineTo",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter tasks with deadline until this date"
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Task"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number",
                          "example": 5
                        },
                        "limit": {
                          "type": "number",
                          "example": 10
                        },
                        "offset": {
                          "type": "number",
                          "example": 0
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "tags": ["Tasks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "404": {
            "description": "List not found"
          },
          "422": {
            "description": "Unprocessable entity - validation errors"
          }
        }
      }
    },
    "/api/tasks/{taskId}": {
      "put": {
        "summary": "Update a task",
        "description": "Updates an existing task with new data",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the task to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "description": "Updated title of the task"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Updated description of the task"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Updated deadline for the task"
                  },
                  "completed": {
                    "type": "boolean",
                    "description": "Updated completion status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/Task"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Deletes an existing task",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the task to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully"
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/lists/{listId}/tasks": {
      "post": {
        "summary": "Create a new task in a specific list",
        "description": "Creates a new task within the specified list (alternative to POST /tasks with listId in body)",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "listId",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID of the list to add the task to"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "description": "Title of the task"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Optional description of the task"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "description": "Optional deadline for the task (future date)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully"
          },
          "400": {
            "description": "Bad request - validation error"
          },
          "404": {
            "description": "List not found"
          }
        }
      }
    },
...
  }
}